# specs/001-build-a-website/contracts/api-specification.yaml
openapi: 3.0.3
info:
  title: Domaine Vallot E-Commerce API
  description: API for boutique winery direct-to-consumer e-commerce website
  version: 1.0.0
  contact:
    name: Domaine Vallot
    email: info@domainevallot.com
servers:
  - url: https://domainevallot.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
paths:
...
  /sync/meta:
    post:
      summary: Sync products to Meta Catalog
      description: Synchronize product data with Facebook/Instagram catalog
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Sync completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  synced_count:
                    type: integer
                  errors:
                    type: array
                    items:
                      type: string
  /sync/google:
    post:
      summary: Sync products to Google Shopping
      description: Synchronize product data with Google Merchant Center
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Sync completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  synced_count:
                    type: integer
                  errors:
                    type: array
                    items:
                      type: string
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...